// src/app/enums/letter-state.enum.ts
/*
 * @Author: Gilberto López
 * @Date: 2018-04-04 04:56:40
 * @Last Modified by: Gilberto López
 * @Last Modified time: 2018-04-04 04:57:28
 */

/* ––
 * –––– Enum definition
 * –––––––––––––––––––––––––––––––––– */

export enum LetterState {
  Undiscovered = 1,
  Discovered = 2,
  Lost = 3
}